requirements.txt
aioredis==2.0.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.2
decouple==0.0.7
distlib==0.3.8
dnspython==2.5.0
et-xmlfile==1.1.0
eventlet==0.35.1
fastapi==0.73.0
filelock==3.13.1
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanize==4.9.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.4
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
openpyxl==3.1.2
package_name==0.1
packaging==23.2
pandas==2.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyarrow==15.0.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.17.1
tornado==6.4
types-pyOpenSSL==24.0.0.20240130
types-redis==4.6.0.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.15.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
